Fixes after checking models with database diagrams

+++ Domain/Models/Assignment.cs
@@ -5,7 +5,7 @@
         public int IdWorker { get; set; }
         public int IdOrder { get; set; }
         public bool OrderLeader { get; set; }
-        public double? HoursWorked { get; set; }
+        public int? HoursWorked { get; set; }
 
         public Worker Worker { get; set; }
         public Order Order { get; set; }
